requirements_training.txt
# DnCNN Training Requirements
# Install with: pip install -r requirements_training.txt

# PyTorch - Core deep learning framework
# For CPU only:
torch>=1.9.0
torchvision>=0.10.0

# For GPU (CUDA) - uncomment and comment the above:
# torch>=1.9.0+cu111
# torchvision>=0.10.0+cu111
# --extra-index-url https://download.pytorch.org/whl/cu111

# Image processing
numpy>=1.20.0
Pillow>=8.0.0
scipy>=1.7.0

# Utilities
tqdm>=4.60.0

# Optional - for TensorBoard logging
# tensorboard>=2.5.0

# The following are already in the main project requirements:
# scikit-image>=0.18.0
# matplotlib>=3.3.0
# opencv-python>=4.5.0

